Up to SimpleDriverDataSource setup.

+++ interfaces/src/main/java/ar/edu/itba/paw/interfaces/UserDao.java
@@ -3,5 +3,5 @@ package ar.edu.itba.paw.interfaces;
 import ar.edu.itba.paw.model.User;
 
 public interface UserDao {
-	public User findById(int id);
+	public User findById(long id);
 }
